feat(server): extract question handler factory into packages/server (#22503)
