+++ packages/server/src/api/index.ts
@@ -1,1 +1,2 @@
-export {}
+export { makeQuestionHandler } from "./question.js"
+export type { QuestionOps } from "./question.js"
